fastapi_app/pyproject.toml
[tool.poetry]
name = "fastapi_app"
version = "0.1.0"
description = ""
authors = ["wellflat <wellflat@gmail.com>"]
packages = [
    {include="src", from="."}
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.74.0"
uvicorn = "^0.17.5"
Pillow = "^9.0.1"
python-multipart = "^0.0.5"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[tool.poetry.scripts]
server = "src.inference.app:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
